WIP pre-precommit
