+++ src/spg/agent/sensor/aerial.py
@@ -1,0 +1,9 @@
+class TopDownSensor(ExternalSensor, ABC):
+
+    """
+    Base class for Image Based sensors.
+    Image based sensors are computed using the top-down rendering of the playground.
+
+    """
+
+
