+++ src/spg/agent/sensor/area.py
@@ -1,0 +1,8 @@
+class AreaSensor(ExternalSensor, ABC):
+    """
+    Base class for Area sensor.
+    Area sensors use Pymunk shape detection to detect shapes within a certain radius.
+    """
+
+
+
